expose useAnswers
