+++ packages/react-instantsearch-dom/src/index.js
@@ -68,6 +68,9 @@ export { default as VoiceSearch } from './widgets/VoiceSearch';
 export { default as QueryRuleCustomData } from './widgets/QueryRuleCustomData';
 export { default as EXPERIMENTAL_Answers } from './widgets/Answers';
 
+// hooks
+export { default as useAnswers } from './hooks/useAnswers';
+
 // Utils
 export { createClassNames } from './core/utils';
